refactor: supabase/proxy.ts를 session.ts로 이름 변경
